@@ -2,7 +2,7 @@
 
   <PropertyGroup>
     <OutputType>Exe</OutputType>
-    <TargetFramework>netcoreapp3.1</TargetFramework>
+    <TargetFramework>net5.0</TargetFramework>
     <RootNamespace>SevenZip.Compression.LZMA</RootNamespace>
     <AssemblyName>TesterApp</AssemblyName>
   </PropertyGroup>
