@@ -1,8 +1,9 @@
 ﻿using System;
 using System.Collections.Generic;
-using Faemiyah.BtDamageResolver.ActorInterfaces;
+using System.Linq;
+using System.Threading.Tasks;
+using Faemiyah.BtDamageResolver.Api.ClientInterface.Events;
 using Faemiyah.BtDamageResolver.Api.Entities;
-using Faemiyah.BtDamageResolver.Api.Events;
 using Microsoft.Extensions.Logging;
 using Orleans;
 
@@ -14,75 +15,59 @@ namespace Faemiyah.BtDamageResolver.Actors
         /// Send all damage reports which have been recorded in this game to the designated player.
         /// </summary>
         /// <param name="playerAuthenticationToken">The authentication token of the player to send the damage reports to.</param>
-        private void DistributeAllDamageReportsToPlayer(Guid playerAuthenticationToken)
+        private async Task DistributeAllDamageReportsToPlayer(Guid playerAuthenticationToken)
         {
-            GrainFactory
-                .GetGrain<IPlayerActor>(_gameActorState.State.AuthenticationTokens[playerAuthenticationToken])
-                .SendDamageReportsToClient(playerAuthenticationToken, _gameActorStateEthereal.DamageReports)
-                .Ignore();
+            await _communicationServiceClient.Send(_gameActorState.State.AuthenticationTokens[playerAuthenticationToken], EventNames.DamageReports, _gameActorState.State.DamageReports.GetAll());
         }
 
         /// <summary>
-        /// Receives a set of damage reports and sends them to players owning the units they concern.
+        /// Receives a set of damage reports and sends them to players.
         /// </summary>
         /// <param name="damageReports">A list of <see cref="DamageReport"/>s that are to be distributed to players.</param>
-        private void DistributeDamageReportsToPlayers(List<DamageReport> damageReports)
+        private async Task DistributeDamageReportsToPlayers(List<DamageReport> damageReports)
         {
-            foreach (var playerId in _gameActorState.State.PlayerStates.Keys)
-            {
-                GrainFactory.GetGrain<IPlayerActor>(playerId).SendDamageReportsToClient(GetAuthenticationTokenForPlayer(playerId), damageReports).Ignore();
-            }
+            _logger.LogInformation("Game {id} is sending an damage reports to all players.", this.GetPrimaryKeyString());
+
+            await _communicationServiceClient.SendToMany(_gameActorState.State.PlayerStates.Keys.ToList(), EventNames.DamageReports, damageReports);
         }
 
         /// <summary>
         /// Sends the game state update to all players.
         /// </summary>
-        private void DistributeGameStateToPlayers()
+        private async Task DistributeGameStateToPlayers()
         {
             _logger.LogInformation("Game {id} is sending a game state update to all players.", this.GetPrimaryKeyString());
             var gameState = GetGameState();
 
-            foreach (var playerId in gameState.Players.Keys)
-            {
-                DistributeGameStateToPlayer(playerId, gameState);
-            }
+            await _communicationServiceClient.SendToMany(gameState.Players.Keys.ToList(), EventNames.GameState, gameState);
         }
 
-        private void DistributeGameStateToPlayer(string playerId, GameState gameState)
+        /// <summary>
+        /// Sends the game state to a single player.
+        /// </summary>
+        private async Task DistributeGameStateToPlayer(Guid authenticationToken)
         {
-            GrainFactory.GetGrain<IPlayerActor>(playerId).SendGameStateToClient(GetAuthenticationTokenForPlayer(playerId), gameState).Ignore();
-        }
-
-        private void DistributeGameStateToPlayer(Guid authenticationToken)
-        {
-            GrainFactory.GetGrain<IPlayerActor>(GetPlayerForAuthenticationToken(authenticationToken)).SendGameStateToClient(authenticationToken, GetGameState()).Ignore();
+            await _communicationServiceClient.Send(GetPlayerForAuthenticationToken(authenticationToken), EventNames.GameState, GetGameState());
         }
 
         /// <summary>
         /// Sends the game options to all players.
         /// </summary>
-        /// <returns>Nothing.</returns>
-        private void DistributeGameOptionsToPlayers()
+        private async Task DistributeGameOptionsToPlayers()
         {
             _logger.LogInformation("Game {id} is sending an options update to all players.", this.GetPrimaryKeyString());
 
-            foreach (var playerId in _gameActorState.State.PlayerStates.Keys)
-            {
-                GrainFactory.GetGrain<IPlayerActor>(playerId).SendGameOptionsToClient(GetAuthenticationTokenForPlayer(playerId), _gameActorState.State.Options).Ignore();
-            }
+            await _communicationServiceClient.SendToMany(_gameActorState.State.PlayerStates.Keys.ToList(), EventNames.GameOptions, _gameActorState.State.Options);
         }
 
         /// <summary>
         /// Sends the target number updates to all players.
         /// </summary>
-        private void DistributeTargetNumberUpdatesToPlayers(List<TargetNumberUpdate> targetNumberUpdates)
+        private async Task DistributeTargetNumberUpdatesToPlayers(List<TargetNumberUpdate> targetNumberUpdates)
         {
             _logger.LogInformation("Game {id} is sending target number updates to all players.", this.GetPrimaryKeyString());
 
-            foreach (var playerId in _gameActorState.State.PlayerStates.Keys)
-            {
-                GrainFactory.GetGrain<IPlayerActor>(playerId).SendTargetNumberUpdatesToClient(GetAuthenticationTokenForPlayer(playerId), targetNumberUpdates).Ignore();
-            }
+            await _communicationServiceClient.SendToMany(_gameActorState.State.PlayerStates.Keys.ToList(), EventNames.TargetNumbers, targetNumberUpdates);
         }
     }
 }
