@@ -1,7 +1,6 @@
 ﻿using System.Collections.Generic;
 using System.Threading.Tasks;
 using Faemiyah.BtDamageResolver.Api.Entities;
-using Faemiyah.BtDamageResolver.Api.Events;
 using Faemiyah.BtDamageResolver.Api.Options;
 
 namespace Faemiyah.BtDamageResolver.Actors.Logic.Interfaces
@@ -10,6 +9,6 @@ namespace Faemiyah.BtDamageResolver.Actors.Logic.Interfaces
     {
         Task<List<DamageReport>> Fire(GameOptions gameOptions, UnitEntry firingUnit);
 
-        Task<DamageReport> ResolveDamageRequest(DamageRequest damageRequest, GameOptions gameOptions);
+        Task<DamageReport> ResolveDamageInstance(DamageInstance damageInstance, GameOptions gameOptions);
     }
 }
