@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Threading.Tasks;
 using Faemiyah.BtDamageResolver.ActorInterfaces;
+using Faemiyah.BtDamageResolver.Api.ClientInterface.Events;
 using Faemiyah.BtDamageResolver.Api.Options;
 using Microsoft.Extensions.Logging;
 using Orleans;
@@ -10,11 +11,11 @@ namespace Faemiyah.BtDamageResolver.Actors
     public partial class PlayerActor
     {
         /// <inheritdoc />
-        public async Task<GameOptions> GetGameOptions(Guid authenticationToken)
+        public async Task<bool> GetGameOptions(Guid authenticationToken)
         {
             if (!await CheckAuthentication(authenticationToken))
             {
-                return null;
+                return false;
             }
 
             _logger.LogInformation("Player {playerId} requested game options.", this.GetPrimaryKeyString());
@@ -22,14 +23,14 @@ namespace Faemiyah.BtDamageResolver.Actors
             if (_playerActorState.State.GameId != null)
             {
                 var grain = GrainFactory.GetGrain<IGameActor>(_playerActorState.State.GameId);
-                return await grain.GetGameOptions(authenticationToken);
+                await SendDataToClient(EventNames.GameOptions, await grain.GetGameOptions(authenticationToken));
             }
 
-            return null;
+            return true;
         }
 
         /// <inheritdoc />
-        public async Task<bool> SetGameOptions(Guid authenticationToken, GameOptions gameOptions)
+        public async Task<bool> SendGameOptions(Guid authenticationToken, GameOptions gameOptions)
         {
             if (!await CheckAuthentication(authenticationToken))
             {
@@ -47,20 +48,21 @@ namespace Faemiyah.BtDamageResolver.Actors
         }
 
         /// <inheritdoc />
-        public async Task<PlayerOptions> GetPlayerOptions(Guid authenticationToken)
+        public async Task<bool> GetPlayerOptions(Guid authenticationToken)
         {
             if (!await CheckAuthentication(authenticationToken))
             {
-                return null;
+                return false;
             }
             
             _logger.LogInformation("Player {playerId} requested player options.", this.GetPrimaryKeyString());
+            await SendDataToClient(EventNames.PlayerOptions, _playerActorState.State.Options);
 
-            return _playerActorState.State.Options;
+            return true;
         }
 
         /// <inheritdoc />
-        public async Task<bool> SetPlayerOptions(Guid authenticationToken, PlayerOptions playerOptions)
+        public async Task<bool> SendPlayerOptions(Guid authenticationToken, PlayerOptions playerOptions)
         {
             if (!await CheckAuthentication(authenticationToken))
             {
