@@ -1,6 +1,7 @@
-﻿using Faemiyah.BtDamageResolver.Actors.Repositories.Base;
-using Faemiyah.BtDamageResolver.Api.Entities;
-using Faemiyah.BtDamageResolver.Api.Interfaces.Repositories;
+﻿using Faemiyah.BtDamageResolver.ActorInterfaces.Repositories;
+using Faemiyah.BtDamageResolver.Actors.Repositories.Prototypes;
+using Faemiyah.BtDamageResolver.Api.ClientInterface.Repositories;
+using Faemiyah.BtDamageResolver.Api.Entities.RepositoryEntities;
 using Microsoft.Extensions.Logging;
 
 namespace Faemiyah.BtDamageResolver.Actors.Repositories
