@@ -1,6 +1,5 @@
 ﻿@using Faemiyah.BtDamageResolver.Client.BlazorServer.Logic
 @inject UserStateController _userStateController
-@implements IDisposable
 
 <div class="resolver_div_componentlistcontainer resolver_div_componentrow">
     <h3 class="resolver_header @(_userStateController.GameState.Players[PlayerId].IsReady ? "ready" : string.Empty)">
@@ -9,21 +8,11 @@
 
     @foreach (var unit in _userStateController.GameState.Players[PlayerId].UnitEntries)
     {
-        <ComponentUnit UnitEntry=@unit></ComponentUnit> // Removed key from here
+        <ComponentUnit UnitEntry=@unit></ComponentUnit> // @key=@($"{_userStateController.PlayerName}_{unit.TimeStamp}")
     }
 </div>
 
 @code {
     [Parameter]
     public string PlayerId { get; set; }
-
-    protected override void OnInitialized()
-    {
-        _userStateController.OnUnitListChange += StateHasChanged;
-    }
-
-    public void Dispose()
-    {
-        _userStateController.OnUnitListChange -= StateHasChanged;
-    }
 }
