@@ -5,7 +5,7 @@
 @inject UserStateController _userStateController
 
 @{
-    var targetNumberUpdate = _userStateController.GetTargetNumberUpdate(WeaponEntry.Id);
+    var targetNumberUpdate = _userStateController.GetTargetNumber(WeaponEntry.Id);
     var targetNumber = targetNumberUpdate?.TargetNumber <= 12 ? $"{targetNumberUpdate.TargetNumber}" : "N/A";
 }
 
