@@ -1,4 +1,5 @@
-﻿@using Faemiyah.BtDamageResolver.Client.BlazorServer.Entities
+﻿@using Microsoft.AspNetCore.Components
+@using Faemiyah.BtDamageResolver.Api.Entities
 
 <div class="resolver_div_componentrow">
     <EditForm Model="@_credentials" OnValidSubmit="Submit">
