@@ -1,4 +1,4 @@
-﻿@using Faemiyah.BtDamageResolver.Api.Entities
+﻿@using Faemiyah.BtDamageResolver.Api.Entities.RepositoryEntities
 @using Faemiyah.BtDamageResolver.Client.BlazorServer.Logic
 @inject CommonData _commonData
 
