@@ -1,4 +1,4 @@
-﻿@using Faemiyah.BtDamageResolver.Api.Entities
+﻿@using Faemiyah.BtDamageResolver.Api.Entities.RepositoryEntities
 @using Faemiyah.BtDamageResolver.Client.BlazorServer.Enums
 @using Faemiyah.BtDamageResolver.Client.BlazorServer.Extensions
 @using Faemiyah.BtDamageResolver.Client.BlazorServer.Logic
@@ -10,6 +10,9 @@
     </td>
     <td>
         @Game.Players
+    </td>
+    <td>
+        @(Game.PasswordProtected ? "Yes" : "No")
     </td>
     <td>
         @{
@@ -35,10 +38,7 @@
         }
     </td>
     <td>
-        @if (!(_userStateController.ConnectedToGame && _userStateController.GameState.GameId == Game.Name))
-        {
-            <button class="resolver_button" @onclick="OnGameJoin">Join</button>
-        }
+        <button class="resolver_button" @onclick="OnGameJoin">Join</button>
     </td>
 </tr>
 
