@@ -1,4 +1,8 @@
-﻿@using Faemiyah.BtDamageResolver.Api.Entities
+﻿@using Faemiyah.BtDamageResolver.Api.Entities.RepositoryEntities
+
+@{
+    Games.Sort((x, y) => y.TimeStamp.CompareTo(x.TimeStamp));
+}
 
 <table class="resolver_table resolver_table_gamelist">
     <tr>
@@ -9,39 +13,42 @@
             Players
         </th>
         <th>
+            Password
+        </th>
+        <th>
             Last activity
         </th>
         <th>
-            <button class="resolver_button" @onclick="RefreshGameList">Refresh</button>
         </th>
     </tr>
-    @if (Games != null)
+    @if (Games.Any())
     {
         @foreach (var game in Games)
         {
             <FormGameEntry Game="@game" OnJoin="@OnGameJoin"></FormGameEntry>
         }
     }
+    else
+    {
+        <tr>
+            <td colspan="5">
+                No recent games
+            </td>
+        </tr>
+    }
 </table>
 
 
-@code {
+@code
+{
     [Parameter]
     public List<GameEntry> Games { get; set; }
 
     [Parameter]
     public EventCallback<string> OnJoin { get; set; }
 
-    [Parameter]
-    public EventCallback<object> OnRefresh { get; set; }
-
     private void OnGameJoin(string gameName)
     {
         OnJoin.InvokeAsync(gameName);
-    }
-
-    private void RefreshGameList()
-    {
-        OnRefresh.InvokeAsync(new object());
     }
 }
