@@ -1,9 +1,8 @@
 ﻿@using Faemiyah.BtDamageResolver.Api.Entities
 @using Faemiyah.BtDamageResolver.Client.BlazorServer.Communication
-@using Faemiyah.BtDamageResolver.Client.BlazorServer.Entities
 @using Faemiyah.BtDamageResolver.Client.BlazorServer.Logic
-@inject Blazored.LocalStorage.ILocalStorageService _localStorage
-@inject IResolverCommunicator _resolverCommunicator
+@using System.Timers
+@inject ResolverCommunicator _resolverCommunicator
 @inject UserStateController _userStateController
 @inject CommonData _commonData
 @implements IDisposable
@@ -36,27 +35,13 @@
                 <button class="resolver_button resolver_button_leave" @onclick="Disconnect">X</button>
             </div>
         </div>
-        @if (!_userStateController.ConnectedToGame && _games != null)
-        {
-            <div class="resolver_div_componentrow">
-                <h2 class="resolver_header">
-                    Games with recent activity
-                </h2>
-                <FormGameList Games="@_recentGames" OnJoin="@OnGameJoinButtonPressed" OnRefresh="@RefreshGameList"></FormGameList>
-            </div>
 
-            @if (_otherGames.Count > 0)
-            {
-                <div class="resolver_div_componentrow">
-                    <h2 class="resolver_header">
-                        Other games
-                    </h2>
-
-                    <FormGameList Games="@_otherGames" OnJoin="@OnGameJoinButtonPressed" OnRefresh="@RefreshGameList"></FormGameList>
-                </div>
-            }
-        }
-
+        <div class="resolver_div_componentrow">
+            <h2 class="resolver_header">
+                Games with recent activity
+            </h2>
+            <FormGameList Games="@_userStateController.GameEntries.Values.ToList()" OnJoin="@OnGameJoinButtonPressed"></FormGameList>
+        </div>
     }
 </div>
 
@@ -89,128 +74,79 @@
 
 @code
 {
-    private static string _localStorageVariableUserCredentials = "BtDamageResolverUserCredentials";
-
     private bool _showModal = false;
     private string _modalName = string.Empty;
     private string _modalPassword = string.Empty;
-    private List<GameEntry> _recentGames;
-    private List<GameEntry> _otherGames;
-    private List<GameEntry> _games;
 
-    [Parameter]
-    public EventCallback OnServerStateChanged { get; set; }
-
-    protected override async Task OnInitializedAsync()
+    protected override void OnInitialized()
     {
-        await RefreshGameList();
-        _userStateController.OnUnitListChange += StateHasChanged;
+        _userStateController.OnGameEntriesReceived += StateHasChanged;
+        RefreshGameList();
 
-        await base.OnInitializedAsync();
+        base.OnInitialized();
     }
 
-    private async Task Connect(Credentials credentials)
+    private void OnTimerInterval(object sender, ElapsedEventArgs e)
     {
-        var loginState = await _resolverCommunicator.Connect(credentials);
+        RefreshGameList();
+    }
 
-        if (loginState != null)
+    public void Connect(Credentials credentials)
+    {
+        _resolverCommunicator.Connect(credentials);
+    }
+
+    private void Disconnect()
+    {
+        _resolverCommunicator.Disconnect();
+    }
+
+    private void JoinGame(Credentials credentials)
+    {
+        _resolverCommunicator.JoinGame(credentials);
+    }
+
+    public void LeaveGame()
+    {
+        _resolverCommunicator.LeaveGame();
+    }
+
+    private void RefreshGameList()
+    {
+        if (!_userStateController.ConnectedToGame)
         {
-            _userStateController.ConnectedToServer = true;
-            _userStateController.PlayerOptions = await _resolverCommunicator.GetPlayerOptions();
-            _userStateController.PlayerName = credentials.Name;
-            await _localStorage.SetItemAsync(_localStorageVariableUserCredentials, credentials);
+            var games = _commonData.GetGameEntries();
 
-            if (!string.IsNullOrWhiteSpace(loginState.GameId))
-            {
-                await JoinGame(new Credentials { Name = loginState.GameId, Password = loginState.GamePassword });
-            }
+            _userStateController.GameEntries = games.ToDictionary(s => s.GetId(), s => s);
+
+            StateHasChanged();
         }
     }
 
-    private async Task Disconnect()
+    private void OnGameJoinButtonPressed(string gameName)
     {
-        if (await _resolverCommunicator.Disconnect())
+        if (_userStateController.GameEntries.TryGetValue(gameName, out var entry) && entry.PasswordProtected)
         {
-            _userStateController.ConnectedToServer = false;
-            _userStateController.GameOptions = null;
-            _userStateController.GameState = null;
-            _userStateController.PlayerName = null;
-            _userStateController.PlayerOptions = null;
-            _userStateController.DamageReportCollection.Clear();
-            await _localStorage.RemoveItemAsync(_localStorageVariableUserCredentials);
-        }
-    }
-
-    private async Task JoinGame(Credentials credentials)
-    {
-        if (await _resolverCommunicator.JoinGame(credentials))
-        {
-            _userStateController.DamageReportCollection.Clear();
-            _userStateController.GameOptions = await _resolverCommunicator.GetGameOptions();
-            await OnServerStateChanged.InvokeAsync();
+            ModalShow();
         }
         else
         {
-            _modalName = credentials.Name;
-            ModalShow();
+            JoinGame(new Credentials { Name = gameName, Password = string.Empty });
         }
     }
 
-    public async Task LeaveGame()
-    {
-        await _resolverCommunicator.LeaveGame();
-    }
-
-    private async Task RefreshGameList()
-    {
-        _games = await _commonData.GetGameEntries();
-        _recentGames = _games.Where(g => g.TimeStamp >= DateTime.UtcNow - TimeSpan.FromHours(1)).ToList();
-        _otherGames = _games.Where(g => g.TimeStamp < DateTime.UtcNow - TimeSpan.FromHours(1)).ToList();
-    }
-
-    private async Task OnGameJoinButtonPressed(string gameName)
-    {
-        await JoinGame(new Credentials { Name = gameName, Password = string.Empty });
-    }
-
-    public void ModalShow()
-    {
-        _showModal = true;
-    }
+    public void ModalShow() => _showModal = true;
 
     private void ModalCancel() => _showModal = false;
 
-    private async Task ModalCompleted(string password)
+    private void ModalCompleted(string password)
     {
         _showModal = false;
-        await JoinGame(new Credentials { Name = _modalName, Password = password });
-    }
-
-    protected override async Task OnAfterRenderAsync(bool firstRender)
-    {
-        if (firstRender)
-        {
-            if (await _localStorage.ContainKeyAsync(_localStorageVariableUserCredentials))
-            {
-                var userCredentials = await _localStorage.GetItemAsync<Credentials>(_localStorageVariableUserCredentials);
-
-                if (!string.IsNullOrWhiteSpace(userCredentials.Name))
-                {
-                    await Connect(userCredentials);
-                }
-            }
-        }
-
-        await base.OnAfterRenderAsync(firstRender);
-    }
-
-    private async Task SaveUserCredentials(Credentials credentials)
-    {
-        await _localStorage.SetItemAsync(_localStorageVariableUserCredentials, credentials);
+        JoinGame(new Credentials { Name = _modalName, Password = password });
     }
 
     public void Dispose()
     {
-        _userStateController.OnUnitListChange -= StateHasChanged;
+        _userStateController.OnGameEntriesReceived -= StateHasChanged;
     }
 }
