@@ -4,10 +4,17 @@
 // 7Zip LZMA compression by Igor Pavlov http://www.7-zip.org/sdk.html
 
 using System;
-using System.Text;
+using System.Collections.Generic;
 
 namespace SevenZip.Compression.LZMA
 {
+    public class ComplexType
+    {
+        public Guid Uuid { get; set; }
+
+        public Dictionary<string,int> Dict { get; set; }
+    }
+
     class Program
     {
         static void Main(string[] args)
@@ -27,32 +34,37 @@ smoother than a V6, while being considerably less expensive than a V12 engine.
 Racing V8s continue to use the single plane crankshaft because it allows faster
 acceleration and more efficient exhaust system designs.";
 
-            // Convert the text into bytes
-            var dataBytes = Encoding.ASCII.GetBytes(OriginalText);
-            Console.WriteLine("Original data is {0} bytes", dataBytes.Length);
-
             // Compress it
-            var compressed = CompressionHelper.Compress(dataBytes);
+            var compressed = DataHelper.Pack(OriginalText);
             Console.WriteLine("Compressed data is {0} bytes", compressed.Length);
 
             // Decompress it
-            var decompressed = CompressionHelper.Decompress(compressed);
+            var decompressed = DataHelper.Unpack<string>(compressed);
             Console.WriteLine("Decompressed data is {0} bytes", decompressed.Length);
 
-            // Convert it back to text
-            var decompressedText = Encoding.ASCII.GetString(decompressed);
-            Console.WriteLine("Is the decompressed text the same as the original? {0}", decompressedText == OriginalText);
+            Console.WriteLine("Is the decompressed text the same as the original? {0}", decompressed == OriginalText);
 
             // Print it out
             Console.WriteLine("And here is the decompressed text:");
-            Console.WriteLine(decompressedText);
+            Console.WriteLine(decompressed);
 
-            var anotherOriginalText = "Kusipää suutu jo!";
-            var byteVersion = CompressionHelper.CompressString(anotherOriginalText);
-            var stringVersion = CompressionHelper.DecompressString(byteVersion);
-            Console.WriteLine(anotherOriginalText);
-            Console.WriteLine(stringVersion);
-            Console.WriteLine(anotherOriginalText.Equals(stringVersion));
+            var originalComplexType = new ComplexType
+            {
+                Uuid = Guid.NewGuid(),
+                Dict = new Dictionary<string, int>
+                {
+                    { "nakki", 715517 },
+                    { "vahvero", 666 }
+                }
+            };
+
+            var originalComplexTypeCompressed = DataHelper.Pack(originalComplexType);
+            var originalComplexTypeUncompressed = DataHelper.Unpack<ComplexType>(originalComplexTypeCompressed);
+            var originalAsText = System.Text.Json.JsonSerializer.Serialize(originalComplexType);
+            var uncompressedAsText = System.Text.Json.JsonSerializer.Serialize(originalComplexTypeUncompressed);
+            Console.WriteLine(originalAsText);
+            Console.WriteLine(uncompressedAsText);
+            Console.WriteLine(originalAsText.Equals(uncompressedAsText));
         }
     }
 }
